Don't get stuck after a poll fails.
On HTTP level errors, the deferred error handler propagated the
error which caused the 'every' to stop scheduling the job, meaning
the monitor no longer updated. Fix here is to not return anything
from the error handlers as I don't see a reason to pass the error
to the next promise (in practice it's ignored anyway).