(fix) Submit form changes with submit() instead of click()
This resolves timing issues in the acceptance tests. Some
configuration pages on Jenkins uses JavaScript to submit the form.
It seems these submits are not caught by Selenium, allowing test
execution to continue which may interrupd the submit. This lead to
changes to views not being saved during acceptance test execution.

Using submit() instead of click() makes sure that Selenium waits
for the request to be submitted before allowing the test execution
to continue.