Move settings to config page as default #119
This commit allows Jenkins users to specify defaults for build monitor
views when creating the views from the Jenkins UI. End users can still
modify these settings using the widget within build monitor.

This is useful because the defaults do not scale to large numbers of builds
so if a user is creating a build monitor view that they anticipate will
contain a large number of builds they can increase the number of columns
and prevent each end user from having to do that individually.

This is essentially a rebase of @tcolligon's changes in pr #189
and is related to issue #119.